+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Improved performance of `delete-orphaned-files` command.
 - Require verification methods to be valid RFC-3986 URIs.
+- Require `keyId` HTTP signature parameter to be valid RFC-3986 URI.
 
 ### Fixed