+++ apx_sdk/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Rename `JsonSigner` enum variants to `HttpUrl` and `DidUrl`.
 - Parse HTTP verification method as `HttpUrl` when verifying JSON signature.
+- Parse `keyId` HTTP signature parameter as `HttpUrl`.
 
 ## [0.4.0] - 2024-11-07